@@ -2,15 +2,14 @@
 using System.Threading.Tasks;
 using Avalonia.Media.Imaging;
 
-namespace AsyncImageLoader
+namespace AsyncImageLoader; 
+
+public interface IAsyncImageLoader : IDisposable
 {
-    public interface IAsyncImageLoader : IDisposable
-    {
-        /// <summary>
-        ///     Loads image
-        /// </summary>
-        /// <param name="url">Target url</param>
-        /// <returns>Bitmap</returns>
-        public Task<Bitmap?> ProvideImageAsync(string url);
-    }
+    /// <summary>
+    ///     Loads image
+    /// </summary>
+    /// <param name="url">Target url</param>
+    /// <returns>Bitmap</returns>
+    public Task<Bitmap?> ProvideImageAsync(string url);
 }
